Support exporting SwitchBot Meter metrics (temperature, humidity and battery).

+++ lib/src/main/kotlin/net/eplusx/mimosa/lib/switchbot/MeterStatusResponse.kt
@@ -21,7 +21,7 @@ data class MeterStatus(
     val deviceType: String,
     val hubDeviceId: String,
     val version: String,
-    val temperature: Float,
+    val temperature: Double,
     val humidity: Int,
     val battery: Int,
 )
